+++ c-ext/python-zstandard.h
@@ -31,7 +31,7 @@
 
 /* Remember to change the string in zstandard/__init__.py, rust-ext/src/lib.rs,
    and debian/changelog as well */
-#define PYTHON_ZSTANDARD_VERSION "0.18.0.dev0"
+#define PYTHON_ZSTANDARD_VERSION "0.18.0"
 
 typedef enum {
     compressorobj_flush_finish,