+++ tests/test_module_attributes.py
@@ -7,7 +7,7 @@ class TestModuleAttributes(unittest.TestCase):
     def test_version(self):
         self.assertEqual(zstd.ZSTD_VERSION, (1, 5, 2))
 
-        self.assertEqual(zstd.__version__, "0.18.0.dev0")
+        self.assertEqual(zstd.__version__, "0.18.0")
 
     def test_features(self):
         self.assertIsInstance(zstd.backend_features, set)