+++ docs/news.rst
@@ -79,8 +79,8 @@ Other Actions Not Blocking Release
 * API for ensuring max memory ceiling isn't exceeded.
 * Move off nose for testing.
 
-0.18.0 (not yet released)
-=========================
+0.18.0 (released 2022-06-20)
+============================
 
 Changes
 -------